### 2018-04-24 bdfbf8e
PAPER: Update date stamp, quote affiliation metadata
