### 2018-04-24 #696 @ajjackson
I have pushed an update to master which should fix the whedon error. Is there a way to trigger a rebuild?

From the former reviewer list, @tacaswell @jochym @mikapfl have listed relevant topics but I do not know if they have access/familiarity with the electronic structure codes that this is intended to post-process (VASP and GPAW). Test data is provided with the package.

@bjmorgan (former reviewer) is well qualified to look at this but I should disclose that he is my mentor in the Software Sustainability Institute [Fellowship Programme](https://www.software.ac.uk/index.php/fellowship-programme)

@jameskermode and @askhl are both VASP users with a good knowledge of open-source development; @askhl is also a GPAW developer.
