### 2018-04-24 6ae3c90
PAPER: Add TYC affil, DOS ERC Acknowledgement, better citation style
