### 2018-04-24 #696 @ajjackson
(Added to affiliations and acknowledgements, changed citation style to one that isn't broken.)
