### 2018-04-24 4a42159
PAPER: Fix chemical formulae in citations
