+++ paper.bib
@@ -110,7 +110,7 @@ journal = {Phys. Rev. B},
 mendeley-groups = {galore},
 number = {23},
 pages = {235109},
-title = {{Experimental and theoretical study of the electronic structure of HgO and Tl{\$}{\_}2{\$}O{\$}{\_}3{\$}}},
+title = {{Experimental and theoretical study of the electronic structure of HgO and Tl$_2$O$_3$}},
 url = {https://link.aps.org/doi/10.1103/PhysRevB.71.235109},
 volume = {71},
 year = {2005}
@@ -197,7 +197,7 @@ mendeley-groups = {galore},
 month = {may},
 number = {20},
 pages = {13445--13457},
-title = {{Photoemission and STM study of the electronic structure of Nb-doped TiO{\$}{\_}2{\$}}},
+title = {{Photoemission and STM study of the electronic structure of Nb-doped TiO$_2$}},
 url = {https://link.aps.org/doi/10.1103/PhysRevB.61.13445},
 volume = {61},
 year = {2000}
