### 2018-04-24 #696 @ajjackson
(Fixed the chemical symbols in some cited paper titles)
