### 2018-04-24 #696 @katyhuff
@tacaswell @jochym @mikapfl @jameskermode or @askhl : Are any of you available to conduct a review for this paper? @ajjackson has submitted it to JOSS and I am the current editor.

The paper is here: https://github.com/openjournals/joss-papers/blob/joss.00696/joss.00696/10.21105.joss.00696.pdf

From the README:

> Galore is a package which applies Gaussian and Lorentzian broadening to data from ab initio calculations. The two main intended applications are
>
> Gaussian and Lorentzian broadening of electronic density-of-states, with orbital weighting to simulate UPS/XPS/HAXPES measurements.
Application of Lorentzian instrumental broadening to simulated Raman spectra from DFPT calculations.
