### 2018-04-26 #696 @ajjackson
Ah, apologies @askhl I had noticed you being active on ASE/Vasp issues! Not very much of the codebase is Vasp-specific, just some import functions which rely on Pymatgen. I would be happy to send extra sample data if needed; Vasp import just needs the files, while GPAW input needs a working GPAW setup.
