### 2018-05-03 b36f4d6
ENH: Add megabarnes unit option
