### 2018-05-03 68d94f2
FIX: Use correct unit conversion
