### 2018-05-03 d8a045a
plot-cs: Singular barn, 79-char line break
