+++ docs/source/conf.py
@@ -63,7 +63,7 @@ author = 'Scanlon Materials Theory Group'
 # The short X.Y version.
 version = '0.5'
 # The full version, including alpha/beta/rc tags.
-release = '0.5.0'
+release = '0.5.1'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
