+++ setup.py
@@ -17,7 +17,7 @@ if __name__ == "__main__":
 
     setup(
         name='galore',
-        version='0.5.0',
+        version='0.5.1',
         description='Broadening and weighting for simulated spectra',
         long_description="""
     Apply Gaussian and Lorentzian broadening to data from ab initio
