### 2018-05-17 #696 @ajjackson
Perhaps @shyamd would be interested? He has provided some very useful feedback on another Vasp-related package of ours.

@askhl when reviewers are set there is another Issue page with checklists for reviewers to use. The reviewer guidelines are located [here](https://joss.theoj.org/about#reviewer_guidelines) (< 1000 words).

@mikapfl the package does come with some sample data, but of course it would be great if reviewers could test with their own. @katyhuff what is your view on this?
