### 2018-06-11 #773 @whedon
[REVIEW]: Galore: Broadening and weighting for simulation of photoelectron spectroscopy
**Submitting author:** @ajjackson (<a href="http://orcid.org/0000-0001-5272-6530">Adam J. Jackson</a>)
**Repository:** <a href="https://github.com/SMTG-UCL/galore" target ="_blank">https://github.com/SMTG-UCL/galore</a>
**Version:** 0.5.0
