### 2018-06-13 afade36
DOC: Start CONTRIBUTING.rst
