### 2018-06-13 e115f34
DOC: CONTRIBUTING code guidelines, README updates
