### 2018-06-13 822f8c3
Update issue templates
