+++ .github/ISSUE_TEMPLATE/question.md
@@ -1,0 +1,7 @@
+---
+name: Question
+about: Other enquiries
+
+---
+
+
