+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,30 @@
+---
+name: Bug report
+about: Create a report to help us improve
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behavior:
+- If using a command-line program (`galore`, `galore-get-cs`, ...) please give a minimal version of the command which produces the error
+- If using the Python API, please provide a minimal Python script which will produce the error
+- Please provide an example which produces the bug with the data files included in the Galore package if possible
+- If the bug only manifests with one of your data files, please provide the file if possible
+  - If you prefer not to share the file, consider generating a small example file that you are willing to share.
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Images**
+If applicable, add image files to help explain your problem.
+
+**System information:**
+ - OS distribution: [e.g. Ubuntu 14.04]
+ - Python version: [e.g. Anaconda Python 3.5]
+ - Matplotlib version: (you can check this with something like `python -c "import matplotlib; print(matplotlib.__version__)"`)
+
+**Additional context**
+Add any other context about the problem here.
