### 2018-06-13 1bee419
Update issue templates
