+++ .github/ISSUE_TEMPLATE/question---idea.md
@@ -1,0 +1,7 @@
+---
+name: Question / Idea
+about: Other enquiries and proposals
+
+---
+
+
