### 2018-06-13 #773 @shyamd
@arfon
This project is a great addition to a growing python based material science community. It provides a nice clean interface to plotting photoelectric spectroscopy data with real world effects quickly. It has clear guidelines on how to contribute, ask questions, and an effective battery of tests. I recommend this paper for publication in JOSS.

P.S. This review process is amazing! I love it.
