+++ paper.bib
@@ -321,6 +321,7 @@ volume = {28},
 number = {165},
 month = {Sept},
 year = {1914},
+pages = {305--319},
 doi = {10.1080/14786440908635214},
 note = {There are a number of other papers from Rutherford's group in this volume of Phil. Mag. related to the ejection of {$\beta$-rays} (electrons) from samples exposed to {$\gamma$-rays} (photons).}
 }
@@ -439,6 +440,7 @@ year = {2015}
   volume={5},
   number={24},
   year={2015},
+  pages={1501462},
   publisher={Wiley Online Library},
   doi={10.1002/aenm.201501462}
 }
