### 2018-06-15 83bc475
DOC: Add JOSS badge to README
