### 2018-06-15 c88ac4f
DOC: fix invisible formatting error
