+++ docs/source/galore.cli.galore_plot_cs.rst
@@ -1,5 +1,5 @@
 galore\.cli\.galore\_plot\_cs module
-===================================
+====================================
 
 .. automodule:: galore.cli.galore_plot_cs
     :members:
