### 2018-06-15 ec18ee8
DOC: Replace paper compilation instructions with journal link
