### 2018-06-15 #773 @arfon
@shyamd many thanks for your review here ✨

@ajjackson - your paper is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00773 :zap: 🚀:boom:
